Rename files
Rename and move files according to reviews
Remove statis assert
